Made the driver class more detailed
Note about the comments I put. I need the jobRequest and Course class before the program can run properly. I did half of the approval class earlier without realizing, so if you all can separate some methods, it will be good. For login, you can do it in an array of objects if you want and separate the logins for the 3 features if you want.