Update driver.java
I have completed the options for Course Coordinator, where the name and ID will be in one array list and the request in another array list. Try it out. If possible do the methods and also expand the casual staff class (view the array list of the request) and enter their availability and the approval class (view both array lists) and create an array list where they can approve/disapprove the request and the course coordinator can view it next time.